+++ gaiku-common/src/lib.rs
@@ -14,10 +14,10 @@ pub trait Baker {
   fn bake(chunk: &Chunk) -> Option<Mesh>;
 
   // TODO: Creating a string key from the coordinates is not the best solution, enhance this
-  fn index(vertices: &mut HashMap<String, (Vector3<f32>, u16)>, vertex: Vector3<f32>) -> u16 {
+  fn index(vertices: &mut HashMap<String, (Vector3<f32>, usize)>, vertex: Vector3<f32>) -> usize {
     let index = vertices.len();
     let key = format!("{:?}", vertex);
-    vertices.entry(key).or_insert((vertex, index as u16)).1
+    vertices.entry(key).or_insert((vertex, index)).1
   }
 }
